Move template namespace files into directory
