+++ inc/templates/admin-bar.php
@@ -7,8 +7,6 @@
 
 namespace WP_Irving\Templates;
 
-use WP_Irving\Components;
-
 /**
  * Automatically insert the admin bar component.
  *
